requirements.txt
pandas
sklearn
matplotlib
seaborn
numpy
keras
pillow
tqmd